Fix sonar issue

+++ systemadmin/src/main/java/com/ecquaria/cloud/moh/iais/action/MasterCodeDelegator.java
@@ -1054,13 +1054,13 @@ public class MasterCodeDelegator {
             syncMasterCodeList.add(oldMasterCodeDto);
             //create new
             masterCodeDto.setMasterCodeId(null);
-            Float version =  masterCodeDto.getVersion();
+            Double version =  masterCodeDto.getVersion();
             if (StringUtil.isEmpty(version)){
                 masterCodeDto.setVersion(1f);
             }else{
                 //get max version ms
                 MasterCodeDto maxMsDto = masterCodeService.getMaxVersionMsDto(oldMasterCodeDto.getMasterCodeKey());
-                Float maxVersion = maxMsDto.getVersion();
+                Double maxVersion = maxMsDto.getVersion();
                 if (maxVersion != null){
                     masterCodeDto.setVersion(maxMsDto.getVersion() + 1);
                 }else{
@@ -1144,7 +1144,7 @@ public class MasterCodeDelegator {
 
             }
         }
-        masterCodeDto.setVersion(StringUtil.isEmpty(ParamUtil.getString(request, MasterCodeConstants.MASTER_CODE_VERSION_CMC)) ? null : Float.parseFloat(ParamUtil.getString(request, MasterCodeConstants.MASTER_CODE_VERSION_CMC)));
+        masterCodeDto.setVersion(StringUtil.isEmpty(ParamUtil.getString(request, MasterCodeConstants.MASTER_CODE_VERSION_CMC)) ? null : Double.parseDouble(ParamUtil.getString(request, MasterCodeConstants.MASTER_CODE_VERSION_CMC)));
         masterCodeDto.setEffectiveFrom(Formatter.parseDate(ParamUtil.getString(request, MasterCodeConstants.MASTER_CODE_EFFECTIVE_FROM_CMC)));
         masterCodeDto.setEffectiveTo(Formatter.parseDate(ParamUtil.getString(request, MasterCodeConstants.MASTER_CODE_EFFECTIVE_TO_CMC)));
         masterCodeDto.setIsEditable(1);
